feat: extract rss languages to its own module
