+++ rss-core/src/main/java/io/micronaut/rss/RssChannel.java
@@ -15,6 +15,7 @@
  */
 package io.micronaut.rss;
 
+import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
@@ -35,6 +36,10 @@ import java.util.Optional;
  */
 public class RssChannel {
 
+    /**
+     * @deprecated Use {@link io.micronaut.rss.language.RssLanguage#LANGUAGE_CODE_PATTERN}.
+     */
+    @Deprecated
     public static final String RSS_LANGUAGE_CODE_REGEX = "af|sq|eu|be|bg|ca|zh-cn|zh-tw|hr|cs|da|nl|nl-be|nl-nl|en|en-au|en-bz|en-ca|en-ie|en-jm|en-nz|en-ph|en-za|en-tt|en-gb|en-us|en-zw| et|fo|fi|fr|fr-be|fr-ca|fr-fr|fr-lu|fr-mc|fr-ch|gl|gd|de|de-at|de-de|de-li|de-lu|de-ch|el|haw|hu|is|in|ga|it|it-it|it-ch|ja|ko|mk|no|pl|pt|pt-br|pt-pt|ro|ro-mo|ro-ro|ru|ru-mo|ru-ru|sr|sk|sl|es|es-ar|es-bo|es-cl|es-co|es-cr|es-do|es-ec|es-sv|es-gt|es-hn|es-mx|es-ni|es-pa|es-py|es-pe|es-pr|es-es|es-uy|es-ve|sv|sv-fi|sv-se|tr|uk";
 
     @NotBlank
@@ -49,9 +54,9 @@ public class RssChannel {
     @NotNull
     private String description;
 
-    @Pattern(regexp = RSS_LANGUAGE_CODE_REGEX)
+    @Pattern(regexp = io.micronaut.rss.language.RssLanguage.LANGUAGE_CODE_PATTERN)
     @Nullable
-    private RssLanguage language;
+    private io.micronaut.rss.language.RssLanguage language;
 
     @Nullable
     private String copyright;
@@ -131,11 +136,12 @@ public class RssChannel {
      * The language the channel is written in. This allows aggregators to group all Italian language sites, for example, on a single page. A list of allowable values for this element, as provided by Netscape, is here. You may also use values defined by the W3C
      * @return The language the channel is written in.
      */
+    @Deprecated
     public Optional<RssLanguage> getLanguage() {
         if (language == null) {
             return Optional.empty();
         }
-        return Optional.of(language);
+        return RssLanguage.of(language.getLanguageCode());
     }
 
     /**
@@ -446,8 +452,18 @@ public class RssChannel {
     /**
      *
      * @param language The language the channel is written in.
+     * @deprecated Use {@link RssChannel#setLanguage(RssLanguage)} instead;
      */
+    @Deprecated
     public void setLanguage(RssLanguage language) {
+        io.micronaut.rss.language.RssLanguage.of(language.getLanguageCode()).ifPresent(rssLanguage -> this.language = rssLanguage);
+    }
+
+    /**
+     *
+     * @param language The language the channel is written in.
+     */
+    public void setLanguage(io.micronaut.rss.language.RssLanguage language) {
         this.language = language;
     }
 
@@ -520,8 +536,21 @@ public class RssChannel {
          *
          * @param language The language the channel is written in.
          * @return The Builder.
+         * @deprecated Use {@link RssChannel.Builder#language} instead.
          */
+        @Deprecated
         public Builder language(RssLanguage language) {
+            rssChannel.setLanguage(language);
+            return this;
+        }
+
+        /**
+         *
+         * @param language The language the channel is written in.
+         * @return The Builder.
+         */
+        @NonNull
+        public Builder language(io.micronaut.rss.language.RssLanguage language) {
             rssChannel.setLanguage(language);
             return this;
         }
