+++ rss-language/build.gradle
@@ -1,14 +1,12 @@
 dependencies {
     annotationProcessor("io.micronaut:micronaut-inject-java")
     annotationProcessor("io.micronaut.docs:micronaut-docs-asciidoc-config-props:$micronautDocsVersion")
+
     annotationProcessor("io.micronaut:micronaut-validation")
     implementation("io.micronaut:micronaut-validation")
-
-    api project(":rss-language")
-    implementation("io.micronaut:micronaut-inject")
+    api("io.micronaut:micronaut-inject")
 
     testImplementation("org.spockframework:spock-core:${spockVersion}") {
         exclude module:'groovy-all'
     }
-    testImplementation("io.micronaut.test:micronaut-test-spock:$micronautTestVersion")
 }
