+++ settings.gradle
@@ -2,4 +2,5 @@ include 'rss'
 include 'rss-core'
 include 'itunespodcast'
 
+include 'rss-language'
 rootProject.name = 'micronaut-rss'
